requirements.txt
pandas
numpy
matplotlib
scikit-learn
tensorflow
keras
joblib
pyyaml
streamlit
